# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: PeopleMatrix synthetic HR lab API. Assessment details are not published here.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/payroll:
    get:
      operationId: listPayroll
      tags: [hr]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Payroll records visible to this role, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/PayrollRecord" } } } } }
  /v1/leave:
    get:
      operationId: listLeave
      tags: [hr]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Leave records visible to this role, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/LeaveRecord" } } } } }
    post:
      operationId: createLeave
      tags: [hr]
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/LeaveInput" } } } }
      responses:
        "201": { description: Requested, content: { application/json: { schema: { $ref: "#/components/schemas/LeaveRecord" } } } }
  /v1/leave/{id}:
    put:
      operationId: updateLeave
      tags: [hr]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/LeaveDecision" } } } }
      responses:
        "200": { description: Reviewed, content: { application/json: { schema: { $ref: "#/components/schemas/LeaveRecord" } } } }
  /v1/performance:
    get:
      operationId: listPerformance
      tags: [hr]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: query, schema: { type: integer } }
      responses:
        "200": { description: Performance records, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/PerformanceRecord" } } } } }
  /v1/performance/comments:
    post:
      operationId: createPerformanceComment
      tags: [hr]
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PerformanceInput" } } } }
      responses:
        "201": { description: Added, content: { application/json: { schema: { $ref: "#/components/schemas/PerformanceRecord" } } } }
  /v1/manager/team:
    get:
      operationId: getManagerTeam
      tags: [hr]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Direct reports, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/EmployeeSummary" } } } } }
  /v1/reports:
    get:
      operationId: getReports
      tags: [hr]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: HR summary, content: { application/json: { schema: { $ref: "#/components/schemas/ReportSummary" } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    EmployeeSummary:
      type: object
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        jobTitle: { type: string }
        department: { type: string }
        managerId: { type: ["integer", "null"] }
      required: [id, firstName, lastName, email, jobTitle, department, managerId]
...
    PayrollRecord:
      type: object
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        period: { type: string }
        gross: { type: string }
        deductions: { type: string }
        net: { type: string }
        paymentDate: { type: string }
      required: [id, employeeId, period, gross, deductions, net, paymentDate]
    LeaveInput:
      type: object
      properties:
        type: { type: string }
        startsOn: { type: string, format: date }
        endsOn: { type: string, format: date }
        reason: { type: string }
      required: [type, startsOn, endsOn, reason]
    LeaveDecision:
      type: object
      properties: { status: { type: string, enum: [APPROVED, REJECTED] } }
      required: [status]
    LeaveRecord:
      type: object
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        type: { type: string }
        startsOn: { type: string }
        endsOn: { type: string }
        reason: { type: string }
        status: { type: string }
        reviewerId: { type: ["integer", "null"] }
        createdAt: { type: string }
      required: [id, employeeId, type, startsOn, endsOn, reason, status, reviewerId, createdAt]
    PerformanceInput:
      type: object
      properties:
        employeeId: { type: integer }
        period: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
      required: [employeeId, period, rating, comment]
    PerformanceRecord:
      type: object
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        reviewerId: { type: integer }
        period: { type: string }
        rating: { type: integer }
        comment: { type: string }
        createdAt: { type: string }
      required: [id, employeeId, reviewerId, period, rating, comment, createdAt]
    ReportSummary:
      type: object
      properties:
        employeeCount: { type: integer }
        departmentCounts: { type: object, additionalProperties: { type: integer } }
        pendingLeave: { type: integer }
      required: [employeeCount, departmentCounts, pendingLeave]
...