# lib/api-zod/src/generated/types/leaveRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PeopleMatrix synthetic HR lab API. Assessment details are not published here.
 * OpenAPI spec version: 0.2.0
 */

export interface LeaveRecord {
  id: number;
  employeeId: number;
  type: string;
  startsOn: string;
  endsOn: string;
  reason: string;
  status: string;
  /** @nullable */
  reviewerId: number | null;
  createdAt: string;
}