# lib/api-zod/src/generated/types/performanceInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PeopleMatrix synthetic HR lab API. Assessment details are not published here.
 * OpenAPI spec version: 0.2.0
 */

export interface PerformanceInput {
  employeeId: number;
  period: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
}